% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC.tvcure.R
\name{AIC.tvcure}
\alias{AIC.tvcure}
\title{Akaike Information Criterion (AIC) of a \code{tvcure.object}.}
\usage{
\method{AIC}{tvcure}(x, ..., k=2)
}
\arguments{
\item{x}{an object of class \code{\link{tvcure.object}}.}

\item{...}{optionally more fitted objects.}

\item{k}{the penalty per parameter to be used (Default: k=2 for the classical AIC).}
}
\value{
The AIC of the tvcure model in \code{x}.
}
\description{
Akaike Information Criterion (AIC) for the fitted tvcure model in a \code{tvcure.object}.
}
\details{
Provides the AIC of the fitted tvcure model in a given \code{\link{tvcure.object}}.
}
\examples{
require(tvcure)
## data(tvcure_Data)
## fit = tvcure(...)
## AIC(fit)

}
\references{
Lambert, P. and Kreyenfeld, M. (2023). Exogenous time-varying covariates in double additive cure survival model
with application to fertility.
\emph{Journal of the Royal Statistical Society, Series A}, in press.
}
\seealso{
\code{\link{tvcure}}, \code{\link{tvcure.object}}, \code{\link{BIC.tvcure}}
}
\author{
Philippe Lambert \email{p.lambert@uliege.be}
}
