% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIC.tvcure.R
\name{BIC.tvcure}
\alias{BIC.tvcure}
\title{Bayesian Information Criterion (BIC) of a \code{tvcure.object}.}
\usage{
\method{BIC}{tvcure}(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link{tvcure.object}}.}

\item{...}{Optionally more fitted objects.}
}
\value{
The BIC of the fitted tvcure model in \code{x}.
}
\description{
Bayesian Information Criterion (BIC) for the fitted tvcure model in a \code{tvcure.object}.
}
\details{
Bayesian (Schwarz) information criterion in a tvcure object, with a penalty calculated using the total effective degrees of freedom and the total number of observed events, -2log(L) + log(d)*ED.tot, smaller values being preferred during model selection.
}
\examples{
require(tvcure)
## data(tvcure_Data)
## fit = tvcure(...)
## BIC(fit)

}
\references{
Lambert, P. and Kreyenfeld, M. (2024). Exogenous time-varying covariates in double additive cure survival model
with application to fertility. \emph{Journal of the Royal Statistical Society, Series A}, in press.
}
\seealso{
\code{\link{tvcure}}, \code{\link{tvcure.object}}, \code{\link{AIC.tvcure}}
}
\author{
Philippe Lambert \email{p.lambert@uliege.be}
}
