% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levidence.R
\name{levidence}
\alias{levidence}
\title{Log-evidence of a \code{tvcure.object}.}
\usage{
\method{levidence}{tvcure}(x, ...)
}
\arguments{
\item{x}{an object of class \code{\link{tvcure.object}}.}

\item{...}{optionally more fitted objects.}
}
\value{
The log-evidence of the tvcure model in \code{x}.
}
\description{
The log-evidence for the fitted tvcure model in a \code{tvcure.object}.
}
\details{
Provides the log of the evidence of the fitted tvcure model in a given \code{\link{tvcure.object}}.
}
\examples{
require(tvcure)
## data(tvcure_Data)
## fit = tvcure(...)
## levidence(fit)

}
\references{
Lambert, P. and Kreyenfeld, M. (2023). Exogenous time-varying covariates in double additive cure survival model
with application to fertility.
\emph{Journal of the Royal Statistical Society, Series A}, in press.
}
\seealso{
\code{\link{tvcure}}, \code{\link{tvcure.object}}, \code{\link{AIC.tvcure}}, \code{\link{BIC.tvcure}}
}
\author{
Philippe Lambert \email{p.lambert@uliege.be}
}
