% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotpredicttvcure.R
\name{plotpredicttvcure}
\alias{plotpredicttvcure}
\title{Plots predicted tvcure object.}
\usage{
plotpredicttvcure(object, type = "S", xlab = "Time",
  ylab = "Predicted Survival Probability", model = c("ph", "aft"), ...)
}
\arguments{
\item{type}{Unknown.}

\item{xlab}{A label for the x-axis.}

\item{ylab}{A label for the y-axis.}

\item{model}{Model type: either "ph" or "aft".}

\item{...}{Further options to be passed to the plot function.}

\item{Object}{An object produced by the predicttvcure function.}
}
\description{
Plots predicted surival curves from objects generated by the predicttvcure function.
}
