% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRegion.R
\name{plotRegion}
\alias{plotRegion}
\title{Generic function to plot a shaded region around values of a scalar function.}
\usage{
plotRegion(x, mat, 
           add=FALSE,xlim=range(x),ylim=range(mat), 
           lwd=2,xlab="",ylab="",main="",...)
}
\arguments{
\item{x}{n-vector with a grid of x values where the scalar function f(x) is evaluated.}

\item{mat}{(n x 3)-matrix containing on its ith row, the function value at x[i] and the bounds of an interval containing it, (f(x[i]), f.low(x[i]), f.up(x[i])).}

\item{add}{logical indicating if the shaded region should be superposed to an existing plot.}

\item{xlim}{x-limits. (Default: range of x).}

\item{ylim}{y-limits. (Default: range of mat).}

\item{lwd}{line width to plot (x,f(x)). (Default: 2).}

\item{xlab}{x-label. (Default: none).}

\item{ylab}{y-label. (Default: none).}

\item{main}{plot main title. (Default: none)}

\item{...}{additional generic plotting arguments.}
}
\value{
No returned value (in addition to the plot).
}
\description{
Generic function to plot a shaded region around values of a scalar function.
}
\references{
Lambert, P. and Kreyenfeld, M. (2024). Exogenous time-varying covariates in double additive cure survival model
with application to fertility. \emph{Journal of the Royal Statistical Society, Series A}, in press.
}
\author{
Philippe Lambert \email{p.lambert@uliege.be}
}
