% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_tvcure.R
\name{predict_tvcure}
\alias{predict_tvcure}
\title{Predict}
\usage{
predict_tvcure(object, newX, newZ, model = c("ph", "aft"), ...)
}
\arguments{
\item{object}{Model object returned from tvcure function.}

\item{newX}{List of new values for X.}

\item{nexZ}{List of new values for Z.}
}
\description{
Predict
}
