% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.R
\name{s}
\alias{s}
\title{Specification of smooth terms in formulas in the tvcure function.}
\usage{
s(x, ref = NULL)
}
\arguments{
\item{x}{Name of the variable for which an additive term is requested.}

\item{ref}{(Optional) reference value for \code{x} where the additive term is zero.}
}
\value{
The submitted variable for which an additive term is required.
}
\description{
Specification of smooth terms in formulas in the tvcure function.
}
\keyword{internal}
