% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{show_c}
\alias{show_c}
\title{Show Copyright Information}
\usage{
show_c()
}
\value{
No return value, called for its side effects of printing a message.
}
\description{
Displays the copyright information for the package in interactive mode.
}
