% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{show_w}
\alias{show_w}
\title{Show Warranty Disclaimer}
\usage{
show_w()
}
\value{
No return value, called for its side effects of printing a message.
}
\description{
Displays a notice that the program comes with absolutely no warranty.
}
