% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.tvcure.R
\name{summary.tvcure}
\alias{summary.tvcure}
\title{Summary method for tvcure models}
\usage{
\method{summary}{tvcure}(model)
}
\arguments{
\item{model}{A model of class tvcure}
}
\description{
Summarizes the output of tvcure models
}
