% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvpred.R
\name{tvpred}
\alias{tvpred}
\title{Prediction function for tvcure models}
\usage{
tvpred(
  model,
  type = c("basesurv", "spop", "suncure", "uncureprob"),
  newX = NULL,
  newZ = NULL,
  CI = T,
  nsims = 1000
)
}
\arguments{
\item{model}{A \link{tvcure} object.}

\item{type}{A character vector indicating values to predict. Choices include the baseline survivor function (\"basesurv\"), the conditional survivor function for uncured individuals (\"suncure\"), the population survivor function (\"spop\"), or the probability that an observation is at risk (\"uncureprob\").}

\item{newX}{An n x k matrix containing values of X to use, where n is the number of observations and k is the number of variables in the hazard equation. Must contain values for all variables in the hazard equation. Must be included unless type = \"basesurv\." Must be passed as a matrix even if it only contains 1 row.}

\item{newZ}{An n x k matrix containing values of Z to use. Must contain values for all variables in the cure equation equation. Must be included unless type = \"basesurv\." Must be passed as a matrix even if it only contains 1 row.}

\item{CI}{Logical value indicating whether confidence intervals should be estimated by simulation.}

\item{nsims}{Number of simulations for estimating confidence intervals.}
}
\description{
Computes various quantities, including the baseline survivor function, conditional survivor function, population survivor function, and probability of being at risk.
}
