% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvtable.R
\name{tvtable}
\alias{tvtable}
\title{Create tables for tvcure and related models}
\usage{
tvtable(
  ...,
  qi = c("se", "zscore", "pvalue"),
  varlist = NULL,
  digits = 3,
  stars = T,
  siglevel = NULL
)
}
\arguments{
\item{...}{One or more models to include in the table. Models may be of class tvcure, coxph, coxphf, or glm.}

\item{qi}{The quantity of interest to be presented below the coefficients. Can be one of "se" for standard errors, "zscore", or "pvalue".}

\item{varlist}{An optional named list containing variable names for each variable. Variables will be presented in the order of the list.}

\item{digits}{The number of digits to which quantities are rounded}

\item{stars}{Logical value indicating whether to include significance stars. By default, \* p < 0.1, \*\* p < 0.05, \*\*\* p < 0.01.}

\item{siglevel}{If stars = T, specifies significance levels and their associated symbols. See \code{\link[modelsummary]{msummary}} for details}
}
\description{
Returns a table stored as a matrix that can be passed to \link[xtable]{name}.
}
