% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC.tvcure.R
\name{AIC.tvcure}
\alias{AIC.tvcure}
\title{Akaike Information Criterion (AIC) of a tvcure object.}
\usage{
\method{AIC}{tvcure}(object, ..., k=2)
}
\arguments{
\item{object}{A \code{\link{tvcure.object}}.}

\item{...}{Other optional tvcure objects.}

\item{k}{The penalty per parameter to be used. (Default: k=2 for the classical AIC).}
}
\value{
The AIC of the fitted tvcure model in \code{x}.
}
\description{
Akaike Information Criterion (AIC) for the fitted tvcure model in a \code{tvcure.object}.
}
\details{
Akaike information criterion for the fitted model in a tvcure object, with a penalty calculated using the total effective degrees of freedom, -2log(L) + 2*ED.tot, larger values being preferred during model selection.
}
\examples{
require(tvcure)
## Simulated data generation
beta = c(beta0=.4, beta1=-.2, beta2=.15) ; gam = c(gam1=.2, gam2=.2) 
df.raw = simulateTVcureData(n=500, seed=123, beta=beta, gam=gam,
                          RC.dist="exponential",mu.cens=550)$df.raw
## TVcure model fitting
tau.0 = 2.5 ; lambda1.0 = c(285,15) ; lambda2.0 = c(25,1325) ## Optional
model = tvcure(~z1+z2+s(x1)+s(x2), ~z3+z4+s(x3)+s(x4), df=df.raw,
               tau.0=tau.0, lambda1.0=lambda1.0, lambda2.0=lambda2.0)
AIC(model)

}
\references{
Lambert, P. and Kreyenfeld, M. (2024). Exogenous time-varying covariates in double additive cure survival model
with application to fertility. \emph{Journal of the Royal Statistical Society, Series A}, under review.
}
\seealso{
\code{\link{tvcure}}, \code{\link{tvcure.object}}, \code{\link{BIC.tvcure}}
}
\author{
Philippe Lambert \email{p.lambert@uliege.be}
}
