% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC.tvcure.R
\name{AIC.tvcure}
\alias{AIC.tvcure}
\title{Akaike Information Criterion (AIC) of a tvcure object.}
\usage{
\method{AIC}{tvcure}(x, ..., k=2)
}
\arguments{
\item{x}{A \code{\link{tvcure.object}}.}

\item{...}{Other optional tvcure objects.}

\item{k}{The penalty per parameter to be used. (Default: k=2 for the classical AIC).}
}
\value{
The AIC of the fitted tvcure model in \code{x}.
}
\description{
Akaike Information Criterion (AIC) for the fitted tvcure model in a \code{tvcure.object}.
}
\details{
Akaike information criterion for the fitted model in a tvcure object, with a penalty calculated using the total effective degrees of freedom, -2log(L) + 2*ED.tot, larger values being preferred during model selection.
}
\examples{
require(tvcure)
## data(tvcure_Data)
## fit = tvcure(...)
## AIC(fit)

}
\references{
Lambert, P. and Kreyenfeld, M. (2024). Exogenous time-varying covariates in double additive cure survival model
with application to fertility. \emph{Journal of the Royal Statistical Society, Series A}, in press.
}
\seealso{
\code{\link{tvcure}}, \code{\link{tvcure.object}}, \code{\link{BIC.tvcure}}
}
\author{
Philippe Lambert \email{p.lambert@uliege.be}
}
