% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIC.tvcure.R
\name{BIC.tvcure}
\alias{BIC.tvcure}
\title{Bayesian Information Criterion (BIC) of a \code{tvcure.object}.}
\usage{
\method{BIC}{tvcure}(x, ...)
}
\arguments{
\item{x}{an object of class \code{\link{tvcure.object}}.}

\item{...}{optionally more fitted objects.}
}
\value{
The BIC of the tvcure model in \code{x}.
}
\description{
Bayesian Information Criterion (BIC) for the fitted tvcure model in a \code{tvcure.object}.
}
\details{
Provides the BIC of the fitted tvcure model in a given \code{\link{tvcure.object}}.
}
\examples{
require(tvcure)
## data(tvcure_Data)
## fit = tvcure(...)
## BIC(fit)

}
\references{
Lambert, P. and Kreyenfeld, M. (2023). Exogenous time-varying covariates in double additive cure survival model
with application to fertility.
\emph{Journal of the Royal Statistical Society, Series A}, in press.
}
\seealso{
\code{\link{tvcure}}, \code{\link{tvcure.object}}, \code{\link{AIC.tvcure}}
}
\author{
Philippe Lambert \email{p.lambert@uliege.be}
}
