% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIC.tvcure.R
\name{BIC.tvcure}
\alias{BIC.tvcure}
\title{Bayesian Information Criterion (BIC) of a \code{tvcure.object}.}
\usage{
\method{BIC}{tvcure}(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{tvcure.object}}.}

\item{...}{Optionally more fitted objects.}
}
\value{
The BIC of the fitted tvcure model in \code{x}.
}
\description{
Bayesian Information Criterion (BIC) for the fitted tvcure model in a \code{tvcure.object}.
}
\details{
Bayesian (Schwarz) information criterion in a tvcure object, with a penalty calculated using the total effective degrees of freedom and the total number of observed events, -2log(L) + log(d)*ED.tot, smaller values being preferred during model selection.
}
\examples{
require(tvcure)
## Simulated data generation
beta = c(beta0=.4, beta1=-.2, beta2=.15) ; gam = c(gam1=.2, gam2=.2)
data = simulateTVcureData(n=500, seed=123, beta=beta, gam=gam,
                          RC.dist="exponential",mu.cens=550)$rawdata
## TVcure model fitting
tau.0 = 2.7 ; lambda1.0 = c(40,15) ; lambda2.0 = c(25,70) ## Optional
model = tvcure(~z1+z2+s(x1)+s(x2), ~z3+z4+s(x3)+s(x4), data=data,
               tau.0=tau.0, lambda1.0=lambda1.0, lambda2.0=lambda2.0)
BIC(model)

}
\references{
Lambert, P. and Kreyenfeld, M. (2024). Exogenous time-varying covariates in double additive cure survival model
with application to fertility. \emph{Journal of the Royal Statistical Society, Series A}, under review.
}
\seealso{
\code{\link{tvcure}}, \code{\link{tvcure.object}}, \code{\link{AIC.tvcure}}, \code{\link{levidence}}
}
\author{
Philippe Lambert \email{p.lambert@uliege.be}
}
