% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviance.tvcure.R
\name{deviance.tvcure}
\alias{deviance.tvcure}
\title{Deviance of a \code{tvcure.object}.}
\usage{
\method{deviance}{tvcure}(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{tvcure.object}}.}

\item{...}{Optionally more fitted objects.}
}
\value{
The deviance as a numeric value, computed according to the model specified in the input \code{object}.
}
\description{
Deviance for the fitted tvcure model in a \code{tvcure.object}.
}
\details{
Deviance -2log(L/L.hat)
}
\examples{
\donttest{
require(tvcure)
## Simulated data generation
beta = c(beta0=.4, beta1=-.2, beta2=.15) ; gam = c(gam1=.2, gam2=.2)
data = simulateTVcureData(n=500, seed=123, beta=beta, gam=gam,
                          RC.dist="exponential",mu.cens=550)$rawdata
## TVcure model fitting
tau.0 = 2.7 ; lambda1.0 = c(40,15) ; lambda2.0 = c(25,70) ## Optional
model = tvcure(~z1+z2+s(x1)+s(x2), ~z3+z4+s(x3)+s(x4), data=data,
               tau.0=tau.0, lambda1.0=lambda1.0, lambda2.0=lambda2.0)
deviance(model)
}

}
\references{
Lambert, P. and Kreyenfeld, M. (2025).
Time-varying exogenous covariates with frequently changing values in double additive cure survival model: an application to fertility.
\emph{Journal of the Royal Statistical Society, Series A}. <doi:10.1093/jrsssa/qnaf035>
}
\seealso{
\code{\link{tvcure}}, \code{\link{tvcure.object}}, \code{\link{AIC.tvcure}}, \code{\link{BIC.tvcure}}, \code{\link{logEvid}}
}
\author{
Philippe Lambert \email{p.lambert@uliege.be}
}
