% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levidence.R
\name{levidence}
\alias{levidence}
\title{Log-evidence of a tvcure object.}
\usage{
levidence(x, ...)
}
\arguments{
\item{x}{A \code{\link{tvcure.object}}.}

\item{...}{Optionally more tvcure objects.}
}
\value{
The log-evidence of the tvcure model in \code{x}.
}
\description{
The log-evidence of the fitted tvcure model in a \code{tvcure.object}.
}
\details{
Provides the log-evidence (or log-marginal likelihood) of the fitted tvcure model in a given \code{\link{tvcure.object}}, where the evidence is the marginal posterior of the penalty parameters at their selected values.
}
\examples{
require(tvcure)
## data(tvcure_Data)
## fit = tvcure(...)
## levidence(fit)

require(tvcure)
## Simulated data generation
beta = c(beta0=.4, beta1=-.2, beta2=.15) ; gam = c(gam1=.2, gam2=.2) 
df.raw = simulateTVcureData(n=500, seed=123, beta=beta, gam=gam,
                          RC.dist="exponential",mu.cens=550)$df.raw
## TVcure model fitting
model = tvcure(~z1+z2+s(x1)+s(x2), ~z3+z4+s(x3)+s(x4), df=df.raw)
levidence(model)
}
\references{
Lambert, P. and Kreyenfeld, M. (2024). Exogenous time-varying covariates in double additive cure survival model
with application to fertility. \emph{Journal of the Royal Statistical Society, Series A}, in press.
}
\seealso{
\code{\link{tvcure}}, \code{\link{tvcure.object}}, \code{\link{AIC.tvcure}}, \code{\link{BIC.tvcure}}
}
\author{
Philippe Lambert \email{p.lambert@uliege.be}
}
