% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logEvid.R
\name{logEvid}
\alias{logEvid}
\title{Generic function for computing log-evidence}
\usage{
logEvid(object, ...)
}
\arguments{
\item{object}{An object for which log-evidence is to be computed.}

\item{...}{Additional arguments (not used for now).}
}
\description{
This is a generic function returning the log-evidence for the class of the input object.
}
