% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.tvcure.R
\name{plot.tvcure}
\alias{plot.tvcure}
\title{Plot visual information related to a tvcure object.}
\usage{
\method{plot}{tvcure}(x, ngrid=300, ci.level=.95, mfrow=NULL, equal.ylims=TRUE, ...)
}
\arguments{
\item{x}{a \code{\link{tvcure.object}}.}

\item{ngrid}{(optional) number of points used to plot the fitted additive terms. (Default: 300).}

\item{ci.level}{(optional) nominal level for the plotted pointwise credible intervals. (Default: 0.95).}

\item{mfrow}{(optional) window layout to plot the additive terms.}

\item{equal.ylims}{logical indicating if the same y-limits must be used when plotting the fitted additive terms. (Default: TRUE).}

\item{...}{additional generic plotting arguments.}
}
\value{
In addition to the plots, an invisible list generated by the \code{\link{additive.tvcure}} function is returned.
}
\description{
Visualization of the estimated additive terms and of the reference (cumulative) hazard function in a tvcure object.
}
\details{
Plot of the fitted additive terms, as well as of the reference hazard \eqn{f_0(t)} and cumulative hazard \eqn{F_0(t)} functions of the fitted tvcure model in \code{x}.
}
\examples{
require(tvcure)
## data(tvcure_Data)
## fit = tvcure(...)
## plot(fit)

}
\references{
Lambert, P. and Kreyenfeld, M. (2024). Exogenous time-varying covariates in double additive cure survival model
with application to fertility. \emph{Journal of the Royal Statistical Society, Series A}, in press.
}
\seealso{
\code{\link{tvcure}}, \code{\link{tvcure.object}}, \code{\link{print.tvcure}}
}
\author{
Philippe Lambert \email{p.lambert@uliege.be}
}
