% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.tvcure.R
\name{plot.tvcure}
\alias{plot.tvcure}
\title{Plot visual information on a \code{tvcure.object}.}
\usage{
\method{plot}{tvcure}(x,
       ngrid=300, ci.level=.95, mfrow=NULL, equal.ylims=TRUE, ...)
}
\arguments{
\item{x}{a \code{\link{tvcure.object}}.}

\item{ngrid}{(optional) number of points to make the plots for the fitted additive terms (default: 300).}

\item{ci.level}{(optional) nominal level for the plotted pointwise credible intervals (default: x$ci.level).}

\item{mfrow}{(optional) window layout to plot the additive terms for location.}

\item{equal.ylims}{logical indicating if the same y-limits must be used when plotting the fitted additive terms (default: TRUE).}

\item{...}{additional generic plotting arguments.}
}
\value{
In addition to the plots, an invisible list generated by the <additive.tvcure> function is returned.
}
\description{
Visualize the estimated additive terms and the reference cumulative hazard function in a tvcure object.
}
\details{
Plot the fitted additive terms and the cdf governing the dynamic of the baseline cumulative hazard function contained in the \code{\link{tvcure.object}} \code{x}.
}
\examples{
require(tvcure)
## data(tvcure_Data)
## fit = tvcure(...)
## plot(fit)

}
\references{
Lambert, P. and Kreyenfeld, M. (2023). Exogenous time-varying covariates in double additive cure survival model
with application to fertility.
\emph{Journal of the Royal Statistical Society, Series A}, in press.
}
\seealso{
\code{\link{tvcure}}, \code{\link{tvcure.object}}, \code{\link{print.tvcure}}
}
\author{
Philippe Lambert \email{p.lambert@uliege.be}
}
