% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.tvcure.R
\name{plot.tvcure}
\alias{plot.tvcure}
\title{Plot visual information related to a tvcure object.}
\usage{
\method{plot}{tvcure}(x, ngrid=300, ci.level=.95, mfrow=NULL, equal.ylims=TRUE, ...)
}
\arguments{
\item{x}{a \code{\link{tvcure.object}}.}

\item{ngrid}{(optional) number of points used to plot the fitted additive terms. (Default: 300).}

\item{ci.level}{(optional) nominal level for the plotted pointwise credible intervals. (Default: 0.95).}

\item{mfrow}{(optional) window layout to plot the additive terms.}

\item{equal.ylims}{logical indicating if the same y-limits must be used when plotting the fitted additive terms. (Default: TRUE).}

\item{...}{additional generic plotting arguments.}
}
\value{
In addition to the plots, an invisible list generated by the \code{\link{additive.tvcure}} function is returned.
}
\description{
Visualization of the estimated additive terms and of the reference (cumulative) hazard function in a tvcure object.
}
\details{
Plot of the fitted additive terms, as well as of the reference hazard \eqn{f_0(t)} and cumulative hazard \eqn{F_0(t)} functions of the fitted tvcure model in \code{x}.
}
\examples{
require(tvcure)
## Simulated data generation
beta = c(beta0=.4, beta1=-.2, beta2=.15) ; gam = c(gam1=.2, gam2=.2) 
df.raw = simulateTVcureData(n=500, seed=123, beta=beta, gam=gam,
                          RC.dist="exponential",mu.cens=550)$df.raw
## TVcure model fitting
tau.0 = 2.7 ; lambda1.0 = c(40,15) ; lambda2.0 = c(25,70) ## Optional
model = tvcure(~z1+z2+s(x1)+s(x2), ~z3+z4+s(x3)+s(x4), df=df.raw,
               tau.0=tau.0, lambda1.0=lambda1.0, lambda2.0=lambda2.0)
plot(model)

}
\references{
Lambert, P. and Kreyenfeld, M. (2024). Exogenous time-varying covariates in double additive cure survival model
with application to fertility. \emph{Journal of the Royal Statistical Society, Series A}, under review.
}
\seealso{
\code{\link{tvcure}}, \code{\link{tvcure.object}}, \code{\link{print.tvcure}}
}
\author{
Philippe Lambert \email{p.lambert@uliege.be}
}
