% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.tvcure.R
\name{predict.tvcure}
\alias{predict.tvcure}
\title{Prediction method for tvcure objects}
\usage{
\method{predict}{tvcure}(
  model,
  type = c("basesurv", "suncure", "spop", "uncureprob"),
  variable = NULL,
  values = NULL,
  newX = NULL,
  newZ = NULL,
  xlab = "Time",
  ylab = NULL,
  legendtitle = NULL,
  internals = F
)
}
\arguments{
\item{model}{A model of class tvcure}

\item{type}{The desired quantity to compute. "basesurv" returns the baseline survivor function based on the observed values of the parameters. "suncure" returns the conditional baseline survival function, while "spop" returns the population (i.e., unconditional) survival function. "Uncureprob" returns the probability that an observation is susceptible to failure (i.e., is not cured.)}

\item{variable}{A string containing the name of the variable to plot predicted values for}

\item{values}{A vector of values of variable at which survival curves should be plotted}

\item{newX}{A matrix containing values of hazard variables to calculate predictions for}

\item{newZ}{A matrix containing values of cure variables to calcluate predictions for}

\item{xlab}{A label for the x-axis}

\item{ylab}{A label for the y-axis}

\item{legendtitle}{Title for the plot legend}

\item{internals}{A logical value indicating whether the predictions should be returned. If FALSE, only the graph will be returned.}

\item{Logical}{value indicating whether graphs should be printed in black and white.}
}
\description{
Computes and plots predicted survival curves and cure probabilities at specified values of independent variables
}
\details{
Values to compute survival curves and cure probabilities for may be specified in one of two ways. First, specifying the variable and values arguments will compute a separate prediction for all values of the specified variable, with all other variables set to their median. Second, for more complex predictions, matrices containing new values of the hazard and cure variables may be passed using newX and newZ. Values contained in newX and newZ will override any specified variables and values if both are included. By default, the predict function returns a plot of the desired quantity. To access the underlying quantities estimated, set internals = TRUE.
}
