% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction4.R
\name{predict.tvcure}
\alias{predict.tvcure}
\title{Prediction method for tvcure objects
##' Computes and plots predicted survival curves and cure probabilities.}
\usage{
\method{predict}{tvcure}(
  model,
  type = c("basesurv", "suncure", "spop", "uncureprob"),
  variable = NULL,
  values = NULL,
  newX = NULL,
  newZ = NULL,
  CI = T,
  nsims = 1000,
  bw = F,
  xlab = "Time",
  ylab = "Predicted Survival Probability",
  legendtitle = NULL,
  internals = F
)
}
\arguments{
\item{model}{A model of class tvcure}

\item{variable}{A character string indicating a variable to plot}

\item{CI}{Logical value indicating whether predictions and confidence intervals should be estimated using maximum simulated likelihood}

\item{nsims}{The number of simulated coefficient values to estimate when computing confidence intervals.}

\item{xlab}{A label for the x-axis.}

\item{ylab}{A label for the y-axis.}

\item{internals}{A logical value indicating whether the predictions should be returned. If FALSE, only the graph will be returned.}

\item{Logical}{value indicating whether graphs should be printed in black and white.}
}
\description{
Prediction method for tvcure objects
##' Computes and plots predicted survival curves and cure probabilities.
}
