% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.tvcure.R
\name{predict.tvcure}
\alias{predict.tvcure}
\title{Predict method for tvcure model fits}
\usage{
\method{predict}{tvcure}(x, df.new, ci.level=.95, ...)
}
\arguments{
\item{x}{A \code{\link{tvcure.object}}.}

\item{df.new}{An optional data frame in which to look for the 'id' (distinguishing the different units), 'time' and covariate values for which 'predictions' should be made. Time values for a given 'id' should be a series of consecutive integers starting with 1. If \code{df.new} is NULL, then predictions are assumed to concern a single unit with consecutive time values starting with 1.}

\item{ci.level}{Credible level for the reported estimates. (Default: 0.95).}

\item{...}{additional generic arguments.}
}
\value{
A list containing, in addition to the optional \code{df.new} entries, the following elements:
\itemize{
\item{\code{Hp} : \verb{ }}{Matrix containing estimates of the cumulative population hazard \eqn{H_p(t|x)} with its credible interval bounds at time \eqn{t} given the history of covariates.}
\item{\code{lHp} : \verb{ }}{Matrix containing estimates of the log cumulative population hazard \eqn{\log H_p(t|x)} with its standard error and credible interval bounds at time \eqn{t} given the history of covariates.}
\item{\code{se.lHp} : \verb{ }}{Vector containing the standard errors of the estimated log cumulative population hazard at time \eqn{t} given the history of covariates.}
\item{\code{hp} : \verb{ }}{Matrix containing estimates of the population hazard \eqn{h_p(t|x)} with its credible interval bounds at time \eqn{t} given the history of covariates.}
\item{\code{lhp} : \verb{ }}{Matrix containing estimates of the log population hazard \eqn{\log h_p(t|x)} with its standard error and credible interval bounds at time \eqn{t} given the history of covariates.}
\item{\code{se.lhp} : \verb{ }}{Vector containing the standard errors of the estimated log population hazard at time \eqn{t} given the history of covariates.}
\item{\code{Sp} : \verb{ }}{Matrix containing estimates of the population survival fuction \eqn{S_p(t|x)=\exp(-H_p(t|x))} with its credible interval bounds at time \eqn{t} given the history of covariates.}
\item{\code{pcure} : \verb{ }}{Matrix containing estimates of the cure probability \eqn{P(cure|x)} with its credible interval bounds at time \eqn{t} if covariates keep their current values.}
\item{\code{llpcure} : \verb{ }}{Matrix containing estimates of the log-log cure probability \eqn{\log(-\log(P(cure|x)))}, with its standard error and credible interval bounds at time \eqn{t} if covariates keep their current values.}
\item{\code{se.llpcure} : \verb{ }}{Vector containing the standard errors of the estimated log-log cure probability \eqn{\log(-\log P(cure|x))} given the history of covariates.}
}
}
\description{
Predicted values based on a tvcure object.
}
\examples{
require(tvcure)
## Simulated data generation
beta = c(beta0=.4, beta1=-.2, beta2=.15) ; gam = c(gam1=.2, gam2=.2) 
df.raw = simulateTVcureData(n=500, seed=123, beta=beta, gam=gam,
                          RC.dist="exponential",mu.cens=550)$df.raw
## TVcure model fitting
model = tvcure(~z1+z2+s(x1)+s(x2), ~z3+z4+s(x3)+s(x4), df=df.raw)

## Covariate profiles for which 'predicted' values are requested
df.new = subset(df.raw, id==1 | id==4)[,-3] ## Focus on units 1 & 4
pred = predict(model,df.new)

## Visualize the estimated population survival fns for units 1 & 4
## par(mfrow=c(1,2))
with(subset(pred,id==1), plotRegion(time,Sp,main="Id=1",
                              ylim=c(0,1),xlab="t",ylab="Sp(t)"))
with(subset(pred,id==4), plotRegion(time,Sp,main="Id=4",
                              ylim=c(0,1),xlab="t",ylab="Sp(t)"))
}
\references{
Lambert, P. and Kreyenfeld, M. (2024). Exogenous time-varying covariates in double additive cure survival model
with application to fertility. \emph{Journal of the Royal Statistical Society, Series A}, under review.
}
\author{
Philippe Lambert \email{p.lambert@uliege.be}
}
