% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvcure.R
\name{tvcure}
\alias{tvcure}
\title{Estimate cure models.}
\usage{
tvcure(
  survform,
  cureform,
  link = "logit",
  data,
  na.action = na.omit,
  offset = NULL,
  subset = NULL,
  var = T,
  nboot = 100,
  parallel = T,
  brglm = F,
  emmax = 1000,
  eps = 1e-07
)
}
\arguments{
\item{var}{If}

\item{nboot}{Specifies the number of bootstrap samples to draw.}

\item{parallel}{If true, bootstraps will be run in parallel. A progress bar displaying the number of completed boostraps will be displayed. This option requires the user to set up a \link{snow} object and register it using the \link{doSNOW} package (see example below).}

\item{emmax}{Specifies the maximum number of iterations for the EM algorithm.}
}
\description{
Estimate cure models.
}
