% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvcure.R
\name{tvcure}
\alias{tvcure}
\title{Estimate cure models.}
\usage{
tvcure(formula, cureform, offset = NULL, model = c("ph", "aft"), data,
  na.action = na.omit, link = "logit", var = T, firthlogit = F,
  firthcox = FALSE, emmax = 50, eps = 1e-07, nboot = 100,
  parallel = T)
}
\description{
Estimate cure models.
}
