% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvcure.R
\name{tvcure}
\alias{tvcure}
\title{Fit of a tvcure model.}
\usage{
tvcure(formula1, formula2, df,
       baseline=c("S0","F0"), K0=20, pen.order0=2,
       K1=10, pen.order1=2, K2=10, pen.order2=2,
       phi.0=NULL, beta.0=NULL, gamma.0=NULL,
       a.tau=1, b.tau=1e-6, a=1, b=1e-2,
       tau.0=100, tau.min=1, tau.method = c("LPS","LPS2","Schall","grid","none"),
       lambda1.0=NULL, lambda1.min=1, lambda2.0=NULL, lambda2.min=1,
       lambda.method=c("LPS","LPS2","LPS3","Schall","nlminb","none"),
       observed.hessian=TRUE, use.Rfast=TRUE, Wood.test=FALSE,
       ci.level=.95,
       criterion=c("levidence","deviance","lpen","AIC","BIC","gradient"),
       criterion.tol=1e-1, grad.tol=1e-2,
       iterlim=50, iter.verbose=FALSE, verbose=FALSE)
}
\arguments{
\item{formula1}{A formula describing the linear predictor in the long-term (cure) survival (or quantum) submodel.}

\item{formula2}{A formula describing the linear predictor in the short-term (cure) survival (or timing) submodel.}

\item{df}{A data frame for survival data in a counting process format. It should always contain at least the following entries:
\itemize{
\item{\code{id} : \verb{ }}{the <id> of the unit associated to the data in a given line in the data frame.}
\item{\code{time} : \verb{ }}{the integer time at which the observations are reported. For a given unit, it should be a sequence of CONSECUTIVE integers starting at 1 for the first observation.}
\item{\code{event} : \verb{ }}{a sequence of 0-1 event indicators. For the lines corresponding to a given unit, it starts with 0 values concluded by a 0 in case of right-censoring or by a 1 if the event is observed at the end of the follow-up.}
}}

\item{baseline}{Baseline ("S0" or "F0") used to specify the dependence of the cumulative hazard dynamics on covariates (Default: "S0"):
Baseline S0: \eqn{S(t|x) = S_0(t)^{\exp^{\gamma'x}}} ;  Baseline F0: \eqn{F(t|x) = F_0(t)^{\exp{\gamma'x}}}}

\item{K0}{Number of B-splines used to specify \eqn{\log f_0(t)} (Default: 20).}

\item{pen.order0}{Penalty order for the P-splines used to specify \eqn{\log f_0(t)} (Default: 2).}

\item{K1}{Number of P-splines for a given additive term in the long-term (or quantum) survival sumbodel (Default: 10).}

\item{pen.order1}{Penalty order for the P-splines in the long-term survival (or quantum) sumbodel (Default: 2).}

\item{K2}{Number of P-splines for a given additive term in the short-term (or timing) survival sumbodel (Default: 10).}

\item{pen.order2}{Penalty order for the P-splines in the short-term survival (or timing) sumbodel (Default: 2).}

\item{phi.0}{(Optional) vector of length \code{K0} with starting values for the P-spline parameters in \eqn{\log f_0(t)}.}

\item{beta.0}{(Optional) starting value for the regression and spline parameters in the long-term survival (or quantum) submodel.}

\item{gamma.0}{(Optional) starting value for the regression and spline parameters in the short-term survival (or timing) submodel.}

\item{a.tau}{Hyperprior parameter in the \eqn{Gamma(a.tau,b.tau)} prior for the penalty parameter \eqn{\tau} tuning the smoothness of \eqn{\log f_0(t)} (Default: 1.0).}

\item{b.tau}{Hyperprior parameter in the \eqn{Gamma(a.tau,b.tau)} prior for the penalty parameter \eqn{\tau} tuning the smoothness of \eqn{\log f_0(t)} (Default: 1e-6).}

\item{a}{Hyperprior parameter in the \eqn{Gamma(a,b)} priors for the penalty parameters \eqn{\lambda_1} and \eqn{\lambda_2} tuning the smoothness of the additive terms in the long-term (quantum) and short-term (timing) survival submodels. (Default: 1.0).}

\item{b}{Hyperprior parameter in the \eqn{Gamma(a,b)} priors for the penalty parameters \eqn{\lambda_1} and \eqn{\lambda_2} tuning the smoothness of the additive terms in the long-term (quantum) and short-term (timing) survival submodels. (Default: 1e-2).}

\item{tau.0}{Starting value for \eqn{\tau}. (Default: 100).}

\item{tau.min}{Minimal value for the penalty parameter \eqn{\tau}. (Default: 1.0).}

\item{tau.method}{Method used to calculate the posterior mode of \eqn{p(\tau|data)}: "LPS", "LPS2", "Schall" (Fellner-Schall algorithm), "grid" (best choice in a regular grid on the log-scale) or "none" (stick to the initial value tau.0). LPS and LPS2, based on Laplace P-splines, both maximize the marginal posterior of the penalty parameter \eqn{\tau} using a fixed-point method, with LPS relying on the prior calculation of eigenvalues. (Default: "LPS").}

\item{lambda1.0}{(Optional) J1-vector with starting values for the penalty parameters of the additive terms in the long-term survival (or quantum) submodel.}

\item{lambda1.min}{Minimal value for the J1 penalty parameters \eqn{\lambda_1} of the additive terms in the long-term survival (or quantum) submodel. (Default: 1.0).}

\item{lambda2.0}{(Optional) J2-vector with starting values for the penalty parameters of the additive terms in the short-term survival (or timing) submodel.}

\item{lambda2.min}{Minimal value for the J2 penalty parameters \eqn{\lambda_2} of the additive terms in the short-term survival (or timing) submodel. (Default: 1.0).}

\item{lambda.method}{Method used ("LPS", "LPS2", "LPS3", "Schall", "nlminb" or "none") to select the penalty parameters of the additive terms in the long-term survival (or quantum) submodel:
\itemize{
\item{\code{LPS}, \code{LPS2}, or \code{LPS3} : \verb{ }}{based on Laplace P-splines where the marginal posterior of the penalty parameters is maximized using a fixed-point method. LPS is based on the prior calculation of eigenvalues (unlike LPS2) and delivers results of comparable quality to those of nlminb, but much more quickly. LPS3 work sequentially and separately on long- and short-term parameters with potentially convergence issues ;}
\item{\code{Schall} : \verb{ }}{Fellner-Schall method ;}
\item{\code{nlminb} : \verb{ }}{nonlinear maximization of the marginal posterior of the penalty parameters using the R function <nlminb> ;}
\item{\code{none} : \verb{ }}{penalty parameters are set at their initial values.}
}}

\item{observed.hessian}{Logical indicating if a fast approximation of the Hessian matrix based on cross-products is preferred over its expected value. (Default: TRUE).}

\item{use.Rfast}{Logical indicating if matrix functions from the Rfast package should be used to fasten computation. (Default: TRUE).}

\item{Wood.test}{Logical indicating if P-values based on Wood's test (Biometrika 2013) of the significance of additive terms should be preferred over basic Chi-square tests. (Default: FALSE).}

\item{ci.level}{Default value for the levels of the credible intervals. (Default: 0.95).}

\item{criterion}{Criterion used to assess convergence of the estimation procedure (Default: "levidence"):
\itemize{
\item{\code{levidence} : \verb{ }}{log of the evidence, i.e. of the marginal posterior of the penalty parameters at their selected values ;}
\item{\code{deviance} : \verb{ }}{deviance or -2 log(Likelihood) ;}
\item{\code{AIC} : \verb{ }}{Akaike information criterion ;}
\item{\code{BIC} : \verb{ }}{Bayesian (or Schwarz) information criterion ;}
\item{\code{gradient} : \verb{ }}{L2 norm of the gradient of the log of the joint posterior w.r.t. the regression and spline parameters.}
}}

\item{criterion.tol}{Maximum absolute difference between the successive values of the \code{criterion} values (when different from "gradient") to declare convergence. (Default: 1e-1).}

\item{grad.tol}{Tolerance value to declare convergence based on gradient values in an optimization procedure (such as Newton-Raphson). (Default: 1e-2).}

\item{iterlim}{Maximum number of iterations. (Default: 50).}

\item{iter.verbose}{Logical indicating if the values of the convergence criterions should be printed after each iteration. (Default: FALSE).}

\item{verbose}{Logical indicating if additional output based on gradients should be printed at the end of each iteration. (Default: FALSE).}
}
\value{
An object of type \code{\link{tvcure.object}}.
}
\description{
Fit of a double additive cure survival model with exogenous time-varying covariates.
}
\examples{
require(tvcure)
## Simulated data generation
beta = c(beta0=.4, beta1=-.2, beta2=.15) ; gam = c(gam1=.2, gam2=.2)
df.raw = simulateTVcureData(n=500, seed=123, beta=beta, gam=gam,
                          RC.dist="exponential",mu.cens=550)$df.raw
## TVcure model fitting
tau.0 = 2.7 ; lambda1.0 = c(40,15) ; lambda2.0 = c(25,70) ## Optional
model = tvcure(~z1+z2+s(x1)+s(x2), ~z3+z4+s(x3)+s(x4), df=df.raw,
               tau.0=tau.0, lambda1.0=lambda1.0, lambda2.0=lambda2.0)
print(model)
plot(model, pages=1)

}
\references{
Lambert, P. and Kreyenfeld, M. (2023). Exogenous time-varying covariates in double additive cure survival model
with application to fertility.
\emph{Journal of the Royal Statistical Society, Series A}, under review.
}
\author{
Philippe Lambert \email{p.lambert@uliege.be}
}
